update sttr, move everything to utilies only keep sttr and model
