+++ openstereo/modeling/models/sttr/utilities/attention.py
