+++ openstereo/modeling/models/sttr/utilities/context_adjustment_layer.py
