+++ openstereo/modeling/models/sttr/utilities/feat_extractor_backbone.py
