+++ openstereo/modeling/models/sttr/utilities/feat_extractor_backbone_in.py
