+++ openstereo/modeling/models/sttr/utilities/transformer.py
